refactor - move allocated_accounts_data_size logic to its own module for easy and clean removal
